+++ pkg/api/types.go
@@ -407,6 +407,18 @@ func (r *Resource) UnmarshalYAML(value *yaml.Node) error {
 			return fmt.Errorf("%s spec: %w", r.ID(), err)
 		}
 		r.Spec = spec
+	case "PortForward":
+		var spec PortForwardSpec
+		if err := raw.Spec.Decode(&spec); err != nil {
+			return fmt.Errorf("%s spec: %w", r.ID(), err)
+		}
+		r.Spec = spec
+	case "IngressService":
+		var spec IngressServiceSpec
+		if err := raw.Spec.Decode(&spec); err != nil {
+			return fmt.Errorf("%s spec: %w", r.ID(), err)
+		}
+		r.Spec = spec
 	case "IPv4PolicyRoute":
 		var spec IPv4PolicyRouteSpec
 		if err := raw.Spec.Decode(&spec); err != nil {